Support for multiple sub-commands added
